--- src/components/@shared/MountFadeInAnimation/MountFadeInAnimation.type.ts
@@ -1,7 +1,0 @@
-import { SerializedStyles } from '@emotion/react';
-
-type MountFadeInAnimationProps = {
-  customStyle?: SerializedStyles;
-};
-
-export { MountFadeInAnimationProps };
