+++ src/components/index.ts
@@ -1,13 +1,13 @@
 import Button from 'components/atoms/Button/Button';
 import Lottie from 'components/atoms/Lottie/Lottie';
 import FadeInAnimation from 'components/atoms/FadeInAnimation/FadeInAnimation';
+import MountFadeInAnimation from 'components/molecules/MountFadeInAnimation/MountFadeInAnimation';
 import ScrollFadeInAnimation from 'components/molecules/ScrollFadeInAnimation/ScrollFadeInAnimation';
 
 import FloatingButton from 'components/@shared/FloatingButton/FloatingButton';
 import FloatingScrollButton from 'components/@shared/FloatingScrollButton/FloatingScrollButton';
 import MDXImageWrapper from 'components/@shared/MDXImageWrapper/MDXImageWrapper';
 import MDXStyleProvider from 'components/@shared/MDXStyleProvider/MDXStyleProvider';
-import MountFadeInAnimation from 'components/@shared/MountFadeInAnimation/MountFadeInAnimation';
 import UnderlineText from 'components/@shared/UnderlineText/UnderlineText';
 import ToggleList from 'components/@shared/ToggleList/ToggleList';
 import Footer from 'components/Footer/Footer';
@@ -17,12 +17,12 @@ export {
   Button,
   Lottie,
   FadeInAnimation,
+  MountFadeInAnimation,
   ScrollFadeInAnimation,
   FloatingButton,
   FloatingScrollButton,
   MDXImageWrapper,
   MDXStyleProvider,
-  MountFadeInAnimation,
   UnderlineText,
   ToggleList,
   Footer,
