+++ src/components/molecules/MountFadeInAnimation/MountFadeInAnimation.type.ts
@@ -1,0 +1,5 @@
+import { FadeInAnimationProps } from 'components/atoms/FadeInAnimation/FadeInAnimation.type';
+
+type MountFadeInAnimationProps = FadeInAnimationProps;
+
+export { MountFadeInAnimationProps };
